refactor: extract admin experiment projection
